requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
sentence-transformers
scikit-learn==1.5.0
torch==2.3.1
python-dotenv==1.0.1
streamlit==1.36.0